+++ include/anpa/internal/combinators_internal.h
@@ -1,9 +1,8 @@
-#ifndef PARSIMON_INTERNAL_COMBINATORS_INTERNAL_H
-#define PARSIMON_INTERNAL_COMBINATORS_INTERNAL_H
+#ifndef ANPA_INTERNAL_COMBINATORS_INTERNAL_H
+#define ANPA_INTERNAL_COMBINATORS_INTERNAL_H
 
 #include <type_traits>
 #include <utility>
-#include <valgrind/callgrind.h>
 #include "anpa/types.h"
 #include "anpa/monad.h"
 #include "anpa/options.h"
@@ -143,4 +142,4 @@ inline constexpr auto lift_or_rec(State& s, InputIt start_pos, Fn f, Parser p, P
 
 }
 
-#endif // PARSIMON_INTERNAL_COMBINATORS_INTERNAL_H
+#endif // ANPA_INTERNAL_COMBINATORS_INTERNAL_H