+++ include/anpa/types.h
@@ -1,5 +1,5 @@
-#ifndef PARSIMON_TYPES_H
-#define PARSIMON_TYPES_H
+#ifndef ANPA_TYPES_H
+#define ANPA_TYPES_H
 
 #include <type_traits>
 #include <iterator>
@@ -75,4 +75,4 @@ inline constexpr void assert_copyable_mreturn() {
 
 
 
-#endif // PARSIMON_TYPES_H
+#endif // ANPA_TYPES_H