+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.3] - 2026-07-15
+### Changed
+- Remove references to old project name
+- Remove unused valgrind include
+
 ## [0.5.2] - 2026-07-15
 ### Changed
 - Fix compilation errors on newer compilers